Improve some tests (#406)

+++ tests/unit/test_tensor.py
@@ -6643,13 +6643,13 @@ def test_torch_nanmedian_keepdim() -> None:
 def test_torch_nansum() -> None:
     assert torch.nansum(
         BatchedTensor(torch.tensor([[0, 1, 2, 3, 4], [5, 6, 7, 8, float("nan")]]))
-    ).equal(torch.tensor(36))
+    ).equal(torch.tensor(36.0))
 
 
 def test_torch_nansum_dim_1() -> None:
     assert torch.nansum(
         BatchedTensor(torch.tensor([[0, 1, 2, 3, 4], [5, 6, 7, 8, float("nan")]])), dim=1
-    ).equal(torch.tensor([10, 26]))
+    ).equal(torch.tensor([10.0, 26.0]))
 
 
 def test_torch_nansum_keepdim() -> None:
@@ -6657,7 +6657,7 @@ def test_torch_nansum_keepdim() -> None:
         BatchedTensor(torch.tensor([[0, 1, 2, 3, 4], [5, 6, 7, 8, float("nan")]])),
         dim=1,
         keepdim=True,
-    ).equal(torch.tensor([[10], [26]]))
+    ).equal(torch.tensor([[10.0], [26.0]]))
 
 
 ################################
@@ -6666,7 +6666,9 @@ def test_torch_nansum_keepdim() -> None:
 
 
 def test_torch_prod() -> None:
-    assert torch.prod(BatchedTensor(torch.arange(10).view(2, 5))).equal(torch.tensor(0.0))
+    assert torch.prod(BatchedTensor(torch.arange(10, dtype=torch.float).view(2, 5))).equal(
+        torch.tensor(0.0)
+    )
 
 
 def test_torch_prod_dim_1() -> None:
